refactor: share ecs factory across controllers
